Move the coverage hook out of bootstrap.cjs
bootstrap.cjs is a --require preload carried by every js_binary runfiles tree
and every container image layer, and the coverage hook had grown it from 1460
to 5886 bytes. The hook now lives in its own node-bootstrap/coverage.cjs,
which bootstrap.cjs requires only when JS_BINARY__COVERAGE_REPORT is set.
js_binary.bzl adds it to the runfiles through the _generates_coverage_report
gate that already carries coverage.js, so only coverage-enabled test targets
pay for it and bootstrap.cjs is back to 1713 bytes.

Along the way, addressing the rest of the review:

- Use the node: prefix on the native requires, so `require` hacks are less
  likely to interfere. The fs patcher mutates the fs module object in place,
  so require('node:fs') still gets the patched one.
- Name the exit callback onExitCoverageReport so it shows up in stack traces.
- Reflow the comments that wrapped with two words on the second line; the
  extraction removes a nesting level, which helps.
- Say why the child test is skipped on Windows: process.execPath is only
  patched to the node wrapper off Windows, so the child would be bare node
  with no --require and child_lib.js's assertions would hold vacuously.
- Read JS_BINARY__LOG_ERROR and JS_BINARY__LOG_PREFIX from the snapshotted
  env rather than live process.env.
- Document that code run from the program's own 'exit' listeners is reported
  as uncovered: a preload cannot register last, so this is inherent to
  producing the report in-process.
