+++ js/private/node-bootstrap/BUILD.bazel
@@ -10,4 +10,5 @@ write_source_files(
 exports_files([
     "fs.cjs",
     "bootstrap.cjs",
+    "coverage.cjs",
 ])
