+++ js/private/test/image/checksum_test.expected
@@ -1,4 +1,4 @@
-b8a78b968f6b2472f42d7276b7819520867ece33906e15ad012b5b6811f5200a  js/private/test/image/cksum_node.tar
+6a7efa791a6d928ca997e6d1c93c57e2cfecc783b862013efc798bbb038c5daf  js/private/test/image/cksum_node.tar
 70b10220a2c05d87da4271c17c38ded8febc725e20e06f1c3809d2bb02ba4ae7  js/private/test/image/cksum_package_store_3p.tar
 2cb6f678d6eb0b2e9d5e2637f41ae3f192233752f1ea2a55cede2531deec2a64  js/private/test/image/cksum_package_store_1p.tar
 79afa99006aff19460354e72cf2634db693670d09ccc7531fbf27f1f20fe0a5f  js/private/test/image/cksum_node_modules.tar
